+++ frontend/src/components/singles/RightBarEventListElement.jsx
@@ -1,0 +1,11 @@
+const RightBarEventListElement = ({event}) =>{
+  return (
+    <li>
+      {event.name}
+      {event.type}
+      {event.time}
+    </li>
+  )  
+}
+
+export default RightBarEventListElement
